move download into own module
